requirements.txt
paddlepaddle-gpu==2.5.1
opencv-python==4.8.0
numpy==1.25.2
onnxruntime
pyctcdecode
tqdm
kenlm
pillow
albumentations